+++ src/app/ng-bootstrap/ng-bootstrap.component.ts
@@ -1,8 +1,7 @@
-import {TourService} from 'ngx-ui-tour-ng-bootstrap';
+import {TourService, IStepOption} from 'ngx-ui-tour-ng-bootstrap';
 import {Component, Inject, OnInit} from '@angular/core';
 import {default as defaultTemplate} from '!!raw-loader!./examples/default-template.txt';
 import {DELAY_AFTER_NAVIGATION} from '../app.providers';
-import {IStepOption} from 'ngx-ui-tour-ng-bootstrap';
 
 
 @Component({